Support Python 2 using future
Changes to support both Python 2 and 3 suggested by `futurize` tool from
the http://python-future.org/ package.

* Use `kwargs` instead of named arguments with default values
* Import `print_function` from `__future__`
* Import backported `range` from `builtins` provided by `future`
* Wrap `_default_options.items()` in a `list`

The `encoding` keyword argument for `open` is only supported in Python
2.6 and 2.7 if imported from the `io` module.

Change the default value of `_call_return_callbacks['chromeFlags']` to
an empty `list` fixing an error attempting to concatenate a `str` to a
`list`.

State the minimum required Python version in `setup.py`.