Update __init__.py to add option
Added option to allow bottle server to listen on all interfaces.  Bottle allows listening on all interfaces when the host is set to '0.0.0.0' as per documentation:  https://bottlepy.org/docs/dev/deployment.html

Without this change setting the host to '0.0.0.0' via eel options fails because this is not a valid address for the web browser to navigate to in order to access the server.  The added option still allows a single IP address to be specified for the browser that eel launches to use, but passes '0.0.0.0' to the bottle server so that the GUI can be accessed on all IP addresses.  For example, both 'localhost' and one or more LAN IPs.  Default is set so that behavior does not change unless this new options is specifically added and set to True.