Fix routing for custom bottles
PR #203 added the ability for users to pass a custom Bottle instance to
Eel, which was intended to help them add middleware. Unfortunately, the
custom app instance has no routing set up and so fundamentally fails to
support Eel's functionality.

This patch applies the functions that handle routing to any custom
bottles before handing off off to bottle to run the server.

Fix #211