Set JS mimetype (#289)
#49 #181 it seems like these issues weren't ever truly fixed.

It would be nice if Eel set the mimetypes on its own. According to https://github.com/pallets/flask/issues/1045 this bug will happen for all Windows users.

The most common use for this would be to import JS modules. Chrome doesn't allow modules with incorrect mimetypes to be loaded without a registry edit.

mimetypes is a Python standard lib so there shouldn't be much impact.

Cheers!