Off-load pyinstaller config to its hook
PyInstaller now has a custom hook for Eel
(https://github.com/pyinstaller/pyinstaller-hooks-contrib/pull/6) that
tells it how to collect eel.js and our hidden imports
(bottle-websocket), so we don't need our own code to enable this.

We can also define a 'distribution' extra for Eel to additionally
install PyInstaller for the user.