Shut down the eel server less aggressively
At the moment, every time a websocket connection closes, we check 1
second later to see if there any any connections open. If a user opens
and terminates multiple websocket connections, this can mean that one of
the earlier "1 second laters" comes along during a later
shutdown/reconnect and closes the server, despite the user navigating
through the app in a reasonable way.

This patch moves this check to a "spawn later" gevent greenlet that can
be terminated and rescheduled. Every time a websocket is closed, we
cancel any previously-running check and schedule a new one for 1 second
in the future, which should mean that Eel shuts down less often during
valid navigation step/s.

Fixes #248