Add some integration tests
At the moment we have no tests, which is bad. We provide a number of
examples that are intended to help users see working examples in order
to grab what they need and get working. It would be nice to
simultaneously validate that our examples are functioning as well as
provide the first bits of assurance that Eel itself is working, so the
first integration tests will run the example code and make assertions
as needed against those. This feels like the most bang-for-buck per test
at the moment.

We also set a default timeout of 5 seconds on all integration tests for
the time being, as (locally, at least) they shouldn't take too long. On
GitHub CI, we bump this to 30 seconds because things seem to run quite a
bit slower there. But we still don't want them to stall and get stuck
forever if we do something wrong. 30 seconds should be more than
generous - but we could bump it if needed.