Add custom_callback option
This allows us to specify the custom command to launch and passing
any argument to it, including the URL of the server that was just created.
Without this option, there is no way to run a custom command
while also giving it the actual url to launch. Hardcoding the url in
cmdline_args being unfeasable if using port=0 for example.